+++ util/transformComponents.js
@@ -3,15 +3,17 @@ import {
   isPlainObject,
   mapValues,
   assign,
-  omit,
   map,
   size,
   keys,
+  pickBy,
 } from "lodash-es";
+import { sep } from "path";
 import {
   parseTemplateReference,
   stringifyTemplateReference,
 } from "./templateReferenceUtils.js";
+import { evaluateBuilders } from "./buildComponents.js";
 
 // As mentioned in the analysis doc, I need to differentiate between
 // YAML nodes created by compacting type identifiers,
@@ -25,7 +27,144 @@ const TYPE_PREFIX = ":";
 const TYPE_REGISTRY = {
   "Common.Content.AssetReference`1[[Artitas.Template, Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null]]":
     "ar_Template",
+  "Artitas.Core.Utils.Reference`1[[Artitas.Template, Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null]]":
+    "0",
+  "Common.Parcels.Components.PackContentsComponent": "PackContents",
+  "Artitas.Template": "4",
 };
+
+// *Technically* a misnomer, since these parsers are for a Reference<T>,
+// but those are backed by an UnsafeOptional<T> so it's fine!
+//
+// (Narrator: It was not fine.)
+function _buildOptionalParser(parser) {
+  return (data) => parseOptional(data, parser);
+}
+
+function parseOptional(data, parser) {
+  if (data === false || data?.IsSet === false) return false;
+  const value = data?.["&v"] ?? data?.Value;
+  return value ? parser(value) : false;
+}
+
+function buildReferenceParser(parser) {
+  return (data) => parseReference(data, parser);
+}
+
+function parseReference(data, parser) {
+  return parser(data.Value);
+}
+
+const PARSE_INTERCEPTORS = {
+  // Food for thought: Could try parsing other AssetReferences too,
+  // under the same principle!
+  ar_Template: parseTemplateReference,
+  0: buildReferenceParser(parseTemplate),
+  // Defined as a ListComponent<Template>.
+  PackContents: (data) => {
+    const { $content, ...rest } = data;
+    const parsedTemplates = $content?.map(parseTemplate);
+    if (keys(rest).length) return { $content: parsedTemplates, ...rest };
+    return parsedTemplates;
+  },
+};
+
+function _buildOptionalStringifier(stringifier) {
+  return (data) => stringifyOptional(data, stringifier);
+}
+
+function stringifyOptional(data, stringifier) {
+  if (!data) return false;
+  // I may have overengineered a tad, *but* consider this:
+  // The overengineering lets me stringify optionals parsed
+  // by older versions of XenoYAML. :D
+  const value = data["&v"] ?? data.Value ?? data;
+  return value ? { "&v": stringifier(value) } : false;
+}
+
+function buildReferenceStringifier(stringifier) {
+  return (data) => stringifyReference(data, stringifier);
+}
+
+function stringifyReference(data, stringifier) {
+  // Catch references parsed by older XenoYAML parser.
+  const value = data.Value ?? data;
+  return { Value: value ? stringifier(value) : null };
+}
+
+const STRINGIFY_INTERCEPTORS = {
+  ar_Template: stringifyTemplateReference,
+  0: buildReferenceStringifier(stringifyTemplate),
+  PackContents: (data) => {
+    if (isArray(data)) return { $content: data.map(stringifyTemplate) };
+    if (isPlainObject(data)) {
+      const { $content, ...rest } = data;
+      const stringifiedTemplates = $content?.map(stringifyTemplate);
+      return { $content: stringifiedTemplates, ...rest };
+    }
+    // Should never happen, but just in case.
+    return data;
+  },
+};
+
+// Extracted template parser and stringifier now that I need them here too.
+export function parseTemplate(data, path) {
+  const {
+    Parent,
+    Name,
+    _components,
+    _excluded,
+    $type: _1,
+    $t: _2,
+    ...rest
+  } = data;
+  const components = parseComponents(_components);
+  const excluded = parseComponents(_excluded);
+  // Catch usage of this function in a map() or similar function.
+  if (path && typeof path !== "string") path = undefined;
+
+  return pickBy({
+    parent: parseTemplateReference(Parent),
+    name: Name,
+    // Portability: Windows understands POSIX path separators,
+    // but most other OSes do not understand Windows separators, so
+    // let's only use POSIX separators in serialized data.
+    ...(path && { $path: path.replaceAll(sep, "/") }),
+    components,
+    excluded,
+    ...rest, // It occurred to me that this could be a thing that happens occasionally.
+  });
+}
+
+export function stringifyTemplate(data, path, builders) {
+  const {
+    parent,
+    name,
+    components,
+    excluded,
+    _components,
+    $path: _,
+    ...rest
+  } = data;
+  // Backwards compatibility check for 0.4.0.
+  if (_components) return stringifyComponent(data);
+  // Catch usage of this function in a map() or similar function.
+  if (path && typeof path !== "string") {
+    path = undefined;
+    builders = undefined;
+  }
+
+  return pickBy({
+    Parent: stringifyTemplateReference(
+      parent,
+      (path || "").replaceAll(sep, "/"),
+    ),
+    Name: name,
+    _components: stringifyComponents(evaluateBuilders(components, builders)),
+    _excluded: stringifyComponents(evaluateBuilders(excluded, builders)),
+    ...rest,
+  });
+}
 
 function parseComponent(data) {
   // 1. Handle Arrays: Keep as arrays, but recurse on children
@@ -41,9 +180,11 @@ function parseComponent(data) {
     if (type) {
       const prefixedType = `${TYPE_PREFIX}${type}`;
 
-      // Special-case template refs.
-      if (type == "ar_Template")
-        return { [prefixedType]: parseTemplateReference(data) };
+      // Special-case handling, for components we
+      // happen to have custom parsing for.
+      if (PARSE_INTERCEPTORS[type]) {
+        return { [prefixedType]: PARSE_INTERCEPTORS[type](rest) };
+      }
 
       const dataKeys = keys(rest);
       if (dataKeys.length === 1 && dataKeys[0] === "$content") {
@@ -59,8 +200,7 @@ function parseComponent(data) {
   return data;
 }
 
-// ... I got tangled up trying to correctly handle inner objects/arrays
-// (the remnants of this failure can still be seen in import.js),
+// ... I got tangled up trying to correctly handle inner objects/arrays,
 // and ended up phoning Gemini for help when my blind eyes couldn't find the problem. :(
 export function parseComponents(data) {
   if (!isArray(data)) return parseComponent(data);
@@ -95,21 +235,15 @@ export function stringifyComponent(data) {
       const originalType = typeKey.substring(TYPE_PREFIX.length);
       const content = data[typeKey];
       const reversedContent = stringifyComponent(content);
+      const registeredType = TYPE_REGISTRY[originalType] ?? originalType;
 
       // Determine if we use $type or $t based on dots
-      const typeProp = originalType.includes(".") ? "$type" : "$t";
+      const typeProp = registeredType.includes(".") ? "$type" : "$t";
 
-      // No sane person would deliberately inject the long form of ar_Template
-      // into their XenoYAML files after I went to the trouble of stripping it out.
-      // Luckily for the insane among us, I'm *juuuust* crazy enough
-      // to conceive of the notion.
-      //
-      // Anyway, since ar_Template has a special parser, it needs a special stringifier...
-      if ((originalType[TYPE_REGISTRY] ?? originalType) == "ar_Template") {
+      if (STRINGIFY_INTERCEPTORS[registeredType]) {
         return {
-          $content: stringifyTemplateReference(content),
-          ...omit(content, ["pack", "screen", "path"]),
-          [typeProp]: originalType,
+          ...STRINGIFY_INTERCEPTORS[registeredType](content),
+          [typeProp]: registeredType,
         };
       }
 
@@ -131,6 +265,13 @@ export function stringifyComponent(data) {
 
 export function stringifyComponents(data) {
   if (!size(data)) return;
+
+  const invalidKeys = keys(data).filter((k) => !k.startsWith(TYPE_PREFIX));
+  if (invalidKeys.length > 0) {
+    throw new SyntaxError(
+      `Found properties without a type prefix (${TYPE_PREFIX}) inside a components block: [${invalidKeys.join(", ")}].`,
+    );
+  }
 
   return map(data, (component, type) => {
     return stringifyComponent({ [type]: component });